Cleaned up code and made some modifications in json & scripts files
